升级rocketmq依赖版本

+++ springboot-rocketmq-producer/src/main/java/com/doodl6/springboot/rocketmq/producer/listener/ClearUserTransactionListener.java
@@ -1,8 +1,10 @@
 package com.doodl6.springboot.rocketmq.producer.listener;
 
+import com.alibaba.fastjson.JSON;
 import com.doodl6.springboot.dao.api.UserMapper;
 import com.doodl6.springboot.dao.entity.User;
 import com.doodl6.springboot.rocketmq.producer.ProducerConstants;
+import com.doodl6.springboot.rocketmq.producer.domain.TransactionMessageObj;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.rocketmq.spring.annotation.RocketMQTransactionListener;
 import org.apache.rocketmq.spring.core.RocketMQLocalTransactionListener;
@@ -18,7 +20,7 @@ import java.nio.charset.Charset;
  */
 @Slf4j
 @Component
-@RocketMQTransactionListener(txProducerGroup = ProducerConstants.TRANSACTION_CLEAR_USER_GROUP)
+@RocketMQTransactionListener
 public class ClearUserTransactionListener implements RocketMQLocalTransactionListener {
 
     private static final String charset = "UTF-8";
@@ -28,14 +30,24 @@ public class ClearUserTransactionListener implements RocketMQLocalTransactionLis
 
     @Override
     public RocketMQLocalTransactionState executeLocalTransaction(Message msg, Object arg) {
-        String msgStr = new String((byte[]) msg.getPayload(), Charset.forName(charset));
-        log.info("收到清除用户执行本地事务消息 | {} | {}", msgStr, arg);
-        Long userId = (Long) arg;
-        int count = userMapper.deleteById(userId);
-        log.info("删除用户信息完成 | userId:{} | count:{}", userId, count);
-
-        //如果返回RocketMQLocalTransactionState.UNKNOWN，则会定时轮训下面的checkLocalTransaction方法检查本地事务状态
-        return RocketMQLocalTransactionState.COMMIT;
+        if (arg instanceof TransactionMessageObj) {
+            TransactionMessageObj obj = (TransactionMessageObj) arg;
+            if (ProducerConstants.TRANSACTION_DELETE_USER.equals(obj.getTransactionType())) {
+                String msgStr = new String((byte[]) msg.getPayload(), Charset.forName(charset));
+                log.info("收到清除用户执行本地事务消息 | msg:{} ", msgStr);
+                Long userId = (Long) obj.getArg();
+                int count = userMapper.deleteById(userId);
+                log.info("删除用户信息完成 | userId:{} | count:{}", userId, count);
+                //如果返回RocketMQLocalTransactionState.UNKNOWN，则会定时轮训下面的checkLocalTransaction方法检查本地事务状态
+                return RocketMQLocalTransactionState.COMMIT;
+            } else {
+                log.warn("收到不支持的本地事务消息 | obj:{}", JSON.toJSONString(obj));
+                return RocketMQLocalTransactionState.UNKNOWN;
+            }
+        } else {
+            log.warn("收到未知的本地事务消息 | arg:{}", JSON.toJSONString(arg));
+            return RocketMQLocalTransactionState.UNKNOWN;
+        }
     }
 
     @Override
